Fixed Scope issue
All the $sevone references failed because they weren't hitting the
global scope.  Turns out none of the functions really worked prior to
this revision... oops.

The issue was masked in my testing as I already had $sevone defined in
the current context.  Anyway, it's fixed now and you are able to import
and run the module.